backend-indexer: refactor exports
