+++ packages/backend-indexer/src/eip1193/indexer/index.ts
@@ -1,3 +1,4 @@
+export * from "./createIndexerEIP1193.js";
 export * from "./requestBlock.js";
 export * from "./requestCode.js";
 export * from "./requestLogs.js";
