+++ packages/backend-indexer/src/eip1193/public/index.ts
@@ -1,0 +1,2 @@
+export * from "./createPublicEIP1193.js"
+export * from "./publicOpenRpcSchema.js"
