+++ packages/backend-indexer/src/eip1193/public/publicOpenRpcSchema.test.ts
